# openapi.yaml
openapi: 3.0.3
info:
  title: WarrantyVault API
  description: |
    REST API for native mobile clients (iOS / Android) of WarrantyVault.
    Source-of-truth for client codegen:
      - iOS:     swift-openapi-generator
      - Android: openapi-generator-cli (kotlin)
    The web app continues to use Next.js Server Actions — those are NOT
    documented here.

    Auth: bearer token from POST /api/v1/auth/login (or /register). Send as
    `Authorization: Bearer <token>` on every authenticated request.

    Ngôn ngữ: mọi chữ trong `message` / `fieldErrors` được server dịch sẵn. Thứ
    tự quyết định: `?lang=vi|en` → header `Accept-Language` → `User.locale` (đã
    lưu) → `en`. Ngôn ngữ không hỗ trợ rơi về mặc định, **không** phải lỗi 400.
    Xem `components.parameters.Lang`.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Local dev
  - url: https://warrantyvault.example.com
    description: Production (set when deployed)
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/backup/export:
    get:
      tags: [backup]
      summary: Download the authenticated user's full data graph (JSON, or .zip kèm ảnh)
      description: |
        Mặc định (`includeBlobs` vắng mặt hoặc `false`): stream **một file JSON**
        chứa devices, warranties, reminders, attachments (metadata + IV /
        wrappedKey base64), wishlist + lịch sử giá, subscriptions + lịch sử thanh
        toán. Phiên bản schema **5** — không đổi so với trước. Response có
        `Content-Disposition: attachment; filename=...`.

        Với `includeBlobs=true`: stream **một file .zip** (`application/zip`,
        filename `.zip`) chứa:
          * `attachments/<storagePath>` — nội dung **đã mã hoá** (AES-256-GCM) của
            từng file đính kèm, đúng byte như trên đĩa (không base64, không nén);
          * `data.json` — envelope phiên bản **6** với
            `includesAttachmentBytes: true` và `attachmentBytesNote` tiếng Việt.
        Nhờ vậy khôi phục sang máy chủ mới mang theo được ảnh hoá đơn — thứ mà bản
        JSON chỉ có metadata không làm được.

        **Sự thật không đổi:** blob là ciphertext, khoá riêng của từng file được
        bọc bằng `FILE_MASTER_KEY`. Khôi phục sang máy chủ thiếu hoặc khác
        `FILE_MASTER_KEY` vẫn ghi lại được dòng dữ liệu và byte, nhưng **không giải
        mã được** — `attachmentBytesNote` nói rõ điều này và kết quả import trả về
        `attachmentsUnreadable`. Đừng hứa với người dùng rằng file backup là tự đủ.

        Nếu một blob không còn trên đĩa lúc export, nó **không** được đưa vào
        archive; id của nó nằm trong `missingAttachmentIds` (và được nhắc trong
        `attachmentBytesNote`) thay vì âm thầm biến mất.
      parameters:
        - name: includeBlobs
          in: query
          required: false
          description: |
            `true` để nhận file `.zip` kèm nội dung ảnh/hoá đơn đã mã hoá thay vì
            JSON chỉ có metadata. Mặc định `false` — hành vi và định dạng JSON y
            như trước. Chỉ nhận `true`/`false` (và `1`/`0`); giá trị khác trả 400.
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: |
            `includeBlobs` vắng/false ⇒ JSON (BackupV1, version 5).
            `includeBlobs=true` ⇒ ZIP: `data.json` (version 6) + các entry
            `attachments/...`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BackupV1' }
            application/zip:
              schema:
                type: string
                format: binary
        '400': { $ref: '#/components/responses/BadInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/backup/import:
    post:
      tags: [backup]
      summary: Restore a backup (JSON metadata, hoặc .zip kèm nội dung ảnh)
      description: |
        Nhận **một trong hai** định dạng, phân biệt bằng magic bytes (`PK\x03\x04`)
        chứ không bằng Content-Type (client hay gửi `application/octet-stream`):

          * **JSON** (`BackupV1`, version 5 hoặc 6) — như trước, chỉ metadata.
            Giới hạn body **50 MB**.
          * **ZIP** do `GET /api/v1/backup/export?includeBlobs=true` tạo ra —
            `data.json` + các entry `attachments/<storagePath>`. Giới hạn body
            **110 MB** (trần upload mỗi người dùng là 100 MB + manifest).

        Import **riêng file `data.json`** của một archive như vậy bị từ chối 400
        kèm hướng dẫn chọn chính file `.zip`: envelope đó khai báo
        `includesAttachmentBytes: true`, nên nhận nó một mình sẽ khôi phục dòng dữ
        liệu mà mất hết ảnh — đúng thứ định dạng blob sinh ra để tránh.

        ZIP được kiểm tra trước khi ghi bất cứ thứ gì: tên entry phải là
        `data.json` hoặc một `storagePath` hợp lệ, mỗi blob ≤ 5 MB, tổng blob
        ≤ 100 MB, `data.json` ≤ 16 MB, và mọi attachment khai báo trong manifest
        phải có entry tương ứng. Ràng buộc của write path được tôn trọng:
        **tối đa 5 file đính kèm / thiết bị** và **tổng 100 MB / người dùng** —
        file bị sửa vượt các ngưỡng này bị từ chối 400 chứ không tạo ra trạng thái
        mà upload thường không cho phép.

        **Id xung đột với tài khoản khác ⇒ 400, không phải 500.** Mọi bảng dùng
        khoá chính text **toàn cục**; file của tài khoản khác sẽ đụng khoá chính.
        Endpoint từ chối trước khi mở transaction, kèm message tiếng Việt nêu rõ
        id và loại thực thể. (Id thuộc chính người đang import được bỏ qua ở chế
        độ `merge` và bị xoá trước ở chế độ `replace`, nên không bị coi là xung
        đột.) Chọn **từ chối** thay vì **đổi id**: đổi id phải viết lại toàn bộ đồ
        thị id (device → warranty → reminder, attachment, wishlist → price,
        subscription → payment) và cả `storagePath` trên đĩa vốn chứa device id —
        sai một cạnh là hỏng âm thầm; đồng thời nó phá tính idempotent của `merge`
        (import lại cùng file sẽ nhân đôi dữ liệu).
      parameters:
        - name: mode
          in: query
          schema: { type: string, enum: [merge, replace], default: merge }
          description: |
            `merge` skips rows whose ids already exist for this user.
            `replace` wipes the user's existing devices / subscriptions /
            wishlist first.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BackupV1' }
          application/zip:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Import succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  result: { $ref: '#/components/schemas/ImportResult' }
        '400': { $ref: '#/components/responses/BadInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413':
          description: Body exceeded the cap (50 MB for JSON, 110 MB for a .zip)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '409':
          description: |
            The payload would break the per-user device ceiling. Importing adds
            device rows, so `services.assertBackupDeviceQuota` refuses a file that
            would push the account past 50 active or 500 stored devices. In
            `replace` mode the base count is zero, so restoring an at-the-limit
            account onto itself is allowed; in `merge` mode ids already owned by
            this user are not counted twice.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /api/v1/cron/warranty-check:
    post:
      tags: [cron]
      security: []
      summary: Run the daily warranty + subscription + wishlist notification sweep
      description: |
        Internal endpoint triggered by a scheduler (systemd timer, Vercel Cron, or
        equivalent). Authenticated via `Authorization: Bearer ${CRON_SECRET}` —
        not the user bearer scheme. Fans out warranty-expiry, wishlist target-date,
        wishlist periodic check-in, and subscription renewal notifications, plus
        auto-bills overdue ACTIVE subscriptions. Idempotent for billing
        (filters by `renewalDate <= now`); see `internal/cron/run.go` for the
        per-bucket idempotency notes.
      parameters:
        - name: Authorization
          in: header
          required: true
          schema: { type: string, example: 'Bearer <CRON_SECRET>' }
      responses:
        '200':
          description: Sweep summary
          content:
            application/json:
              schema:
                type: object
                required: [ok, checkedAt, stats]
                properties:
                  ok: { type: boolean }
                  checkedAt: { type: string, format: date-time }
                  stats:
                    type: object
                    description: |
                      Bộ đếm best-effort của một lượt chạy. Một hàng lỗi bị ghi log
                      và bỏ qua, không làm hỏng phần còn lại của job — nên các con số
                      này là "đã xử lý", không phải "đã thành công".
                    required:
                      - warrantyNotices
                      - returnWindowNotices
                      - wishlistTargetHits
                      - wishlistCheckins
                      - subscriptionRenewals
                      - subscriptionExpired
                      - sessionsPruned
                      - pushesSent
                      - pushesFailed
                      - pushesGone
                    properties:
                      warrantyNotices: { type: integer, description: 'Số gói bảo hành rơi vào mốc 7 hoặc 30 ngày trước `endDate`.' }
                      returnWindowNotices: { type: integer, description: 'Số thiết bị rơi vào mốc còn 3 hoặc 1 ngày đổi/trả (`COALESCE(receivedAt, purchaseDate) + returnWindowDays`).' }
                      wishlistTargetHits: { type: integer }
                      wishlistCheckins: { type: integer }
                      subscriptionRenewals: { type: integer, description: 'Số gói đã được tự động gia hạn THẬT trong một transaction (có ghi `SubscriptionPayment`).' }
                      subscriptionExpired: { type: integer }
                      sessionsPruned: { type: integer }
                      pushesSent: { type: integer }
                      pushesFailed: { type: integer }
                      pushesGone: { type: integer, description: 'Push trả 404/410 — hàng `PushSubscription` tương ứng đã bị xoá.' }
        '401': { description: 'Missing or wrong CRON_SECRET' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  responses:
...
    BadInput:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: Missing / invalid / expired bearer token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error: { type: string, description: 'Stable code, e.g. unauthorized, not_found, bad_input, rate_limited, limit_reached' }
        message: { type: string, description: 'Human-readable Vietnamese message' }
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
...
    BackupV1:
      type: object
      description: |
        Envelope dữ liệu người dùng. Hai định dạng cùng tồn tại:

          * **version 5** — file JSON của `GET /api/v1/backup/export` (mặc định)
            **và** mọi file người dùng đã tải trước đây. Chỉ có metadata của file
            đính kèm, `includesAttachmentBytes: false`.
          * **version 6** — `data.json` **bên trong file .zip** của
            `GET /api/v1/backup/export?includeBlobs=true`, với
            `includesAttachmentBytes: true`; nội dung ảnh nằm ở các entry
            `attachments/<storagePath>` cạnh nó trong archive.

        Round-trip ổn định: export rồi import lại trên tài khoản/database sạch
        (chế độ `replace`) cho cùng tập dòng dữ liệu.

        **Sự thật về file đính kèm.** Bản JSON (v5) KHÔNG chứa nội dung ảnh, nên
        khôi phục sang máy chủ mới sẽ mất ảnh hoá đơn. Bản ZIP (v6) CÓ chứa nội
        dung đã mã hoá, nhưng blob là ciphertext AES-256-GCM với khoá riêng từng
        file được bọc bằng `FILE_MASTER_KEY`: thiếu hoặc sai khoá thì byte vẫn còn
        mà ảnh không mở được. `includesAttachmentBytes` / `attachmentBytesNote`
        nói đúng điều đó ngay trong payload để client cảnh báo chính xác thay vì
        hardcode. Hai trường này là metadata của bản export: import bỏ qua chúng
        khi quyết định, và payload cũ thiếu chúng vẫn import được.

        Import chấp nhận **mọi** phiên bản trong khoảng
        [MinBackupVersion, BackupVersion] (hiện tại 5-6) thay vì so bằng đúng, nên
        việc nâng lên v6 không làm file v5 của người dùng thành rác. Payload mới
        hơn (ví dụ 7) bị từ chối với thông báo tiếng Việt nói rõ phiên bản nhận
        được; payload quá cũ (≤4) cũng vậy.
      required: [version, exportedAt, devices, wishlist, subscriptions]
      properties:
        version:
          type: integer
          enum: [5, 6]
          description: |
            `5` = JSON chỉ metadata (định dạng cũ, vẫn được ghi cho tới khi client
            đổi). `6` = envelope trong file .zip có kèm nội dung ảnh. Import chấp
            nhận cả hai.
        exportedAt: { type: string, format: date-time }
        includesAttachmentBytes:
          type: boolean
          description: |
            `false` ở bản JSON v5: payload KHÔNG chứa nội dung ảnh/hoá đơn, chỉ có
            metadata (`fileName`, `fileType`, `fileSize`, `iv`, `wrappedKey`) và
            `storagePath`. `true` ở `data.json` bên trong file .zip v6: nội dung
            đã mã hoá của từng attachment nằm ở entry
            `attachments/<storagePath>` trong cùng archive.
        attachmentBytesNote:
          type: string
          description: 'Câu tiếng Việt giải thích `includesAttachmentBytes` (và, với bản ZIP, cả yêu cầu `FILE_MASTER_KEY`) để client hiển thị thẳng cho người dùng.'
        missingAttachmentIds:
          type: array
          items: { type: string }
          description: |
            Chỉ có ở bản ZIP: id của những attachment mà **blob không còn trên đĩa**
            lúc export, nên không thể đưa vào archive. Dòng dữ liệu vẫn được export
            và import lại (khôi phục metadata, thiếu byte); trường này để client
            nói thật với người dùng thay vì để họ phát hiện lúc mở hoá đơn. Vắng
            mặt khi mọi blob đều có.
        subscriptions:
          type: array
          items: { $ref: '#/components/schemas/BackupSubscription' }
        wishlist:
          type: array
          items: { $ref: '#/components/schemas/BackupWishlistItem' }
        devices:
          type: array
          items: { $ref: '#/components/schemas/BackupDevice' }
    BackupSubscription:
      type: object
      required: [id, name, billingCycle, price, currency, startedAt, renewalDate, autoRenew, status, createdAt, updatedAt, payments]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string, nullable: true }
        brand: { type: string, nullable: true }
        plan: { type: string, nullable: true }
        billingCycle: { type: string }
        intervalDays: { type: integer, nullable: true }
        price: { type: integer }
        currency: { type: string }
        startedAt: { type: string, format: date-time }
        renewalDate: { type: string, format: date-time }
        autoRenew: { type: boolean }
        status: { type: string }
        accountEmail: { type: string, nullable: true }
        paymentMethod: { type: string, nullable: true }
        manageUrl: { type: string, nullable: true }
        cancelUrl: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        lastNotifiedRenewalAt:
          type: string
          format: date-time
          nullable: true
          description: |
            Mốc chống trùng thông báo gia hạn của cron. Được export/import nguyên
            vẹn: nếu thiếu, lần chạy cron kế tiếp sẽ nhắc lại các mốc 3/1/0 ngày
            mà người dùng đã được nhắc. Null = chưa từng nhắc.
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        payments:
          type: array
          items:
            type: object
            required: [id, amount, paidAt]
            properties:
              id: { type: string }
              amount: { type: integer }
              paidAt: { type: string, format: date-time }
              note: { type: string, nullable: true }
    BackupWishlistItem:
      type: object
      required: [id, name, priority, status, createdAt, updatedAt, prices]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string, nullable: true }
        brand: { type: string, nullable: true }
        initialPrice: { type: integer, nullable: true }
        currentPrice: { type: integer, nullable: true }
        buyUrl: { type: string, nullable: true }
        imageUrl: { type: string, nullable: true }
        targetDate: { type: string, format: date-time, nullable: true }
        priority: { type: string }
        status: { type: string }
        notes: { type: string, nullable: true }
        reminderIntervalDays: { type: integer, nullable: true }
        lastNotifiedAt: { type: string, format: date-time, nullable: true }
        purchasedDeviceId: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        prices:
          type: array
          items:
            type: object
            required: [id, price, recordedAt]
            properties:
              id: { type: string }
              price: { type: integer }
              note: { type: string, nullable: true }
              recordedAt: { type: string, format: date-time }
    BackupDevice:
      type: object
      required: [id, name, category, purchaseDate, purchasePrice, status, createdAt, updatedAt, warranties, attachments]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        brand: { type: string, nullable: true }
        model: { type: string, nullable: true }
        serialNumber: { type: string, nullable: true }
        purchaseDate: { type: string, format: date-time }
        purchasePrice: { type: integer }
        purchasePlace: { type: string, nullable: true }
        status: { type: string }
        notes: { type: string, nullable: true }
        soldAt:
          type: string
          format: date-time
          nullable: true
          description: 'Thời điểm bán (migration 0006). Vắng mặt/null trong file backup cũ = chưa bán, vẫn import được.'
        soldPrice: { type: integer, nullable: true, description: 'Giá bán, VND. Null = chưa bán.' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        warranties:
          type: array
          items:
            type: object
            required: [id, type, startDate, endDate, months, createdAt, updatedAt, reminders]
            properties:
              id: { type: string }
              type: { type: string }
              provider: { type: string, nullable: true }
              startDate: { type: string, format: date-time }
              endDate: { type: string, format: date-time }
              months: { type: integer }
              cost: { type: integer, nullable: true }
              address: { type: string, nullable: true }
              phone: { type: string, nullable: true }
              notes: { type: string, nullable: true }
              createdAt: { type: string, format: date-time }
              updatedAt: { type: string, format: date-time }
              reminders:
                type: array
                items:
                  type: object
                  required: [id, isDismissed, createdAt]
                  properties:
                    id: { type: string }
                    isDismissed: { type: boolean }
                    lastNotifiedAt:
                      type: string
                      format: date-time
                      nullable: true
                      description: |
                        Mốc cron đã gửi nhắc lần cuối (chống trùng, migration
                        0002). Được export/import nguyên vẹn — thiếu nó thì lần
                        cron kế tiếp sẽ nhắc lại bảo hành người dùng đã được nhắc.
                        Null = chưa từng nhắc.
                    createdAt: { type: string, format: date-time }
        attachments:
          type: array
          items:
            type: object
            required: [id, fileName, storagePath, fileType, fileSize, iv, wrappedKey, uploadedAt]
            properties:
              id: { type: string }
              fileName: { type: string }
              storagePath: { type: string, description: '<deviceId>/<uuid>.enc' }
              fileType: { type: string }
              fileSize: { type: integer }
              iv: { type: string, description: 'base64-encoded 12-byte IV' }
              wrappedKey: { type: string, description: 'base64-encoded wrapped per-file key' }
              description: { type: string, nullable: true }
              uploadedAt: { type: string, format: date-time }
    ImportResult:
      type: object
      required: [imported, skipped, wishlistImported, wishlistSkipped, subImported, subSkipped, attachmentsImported, attachmentsSkipped, attachmentsUnreadable]
      properties:
        imported: { type: integer }
        skipped: { type: integer }
        wishlistImported: { type: integer }
        wishlistSkipped: { type: integer }
        subImported: { type: integer }
        subSkipped: { type: integer }
        attachmentsImported:
          type: integer
          description: |
            Số file đính kèm đã ghi ra `PRIVATE_UPLOAD_ROOT` (chỉ khác 0 khi import
            file .zip có kèm nội dung ảnh). Import JSON chỉ metadata luôn trả 0.
        attachmentsSkipped:
          type: integer
          description: |
            Số file đính kèm bị bỏ qua cùng thiết bị của chúng (chế độ `merge`,
            thiết bị đã tồn tại trong tài khoản) — byte cũng không được ghi, nên
            không sinh file mồ côi.
        attachmentsUnreadable:
          type: integer
          description: |
            Số file đính kèm **không dùng được**: hoặc không giải mã được bằng
            `FILE_MASTER_KEY` của máy chủ đang import (kể cả khi máy chủ chưa cấu
            hình khoá — bản backup đến từ máy chủ khác), hoặc manifest đã đánh dấu
            là thiếu blob từ lúc export (`missingAttachmentIds`). Dòng dữ liệu và
            byte vẫn được khôi phục; con số này để client cảnh báo người dùng.
...